2nrv3 sorta working
